pyproject.toml
[project]
name = "navermail-mcp-server"
version = "0.1.0"
description = "Navermail MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imap-tools>=1.11.0",
    "mcp[cli]>=1.12.1",
    "python-dotenv>=1.1.1",
]
